db_text: Cast time_t
CID #417516, #417608, #417659, #417677, #417784

+++ modules/db_text/dbt_file.c
@@ -69,7 +69,7 @@ int dbt_check_mtime(const str *tbn, const str *dbn, time_t *mt)
 	}
 	if(stat(path, &s) == 0)
 	{
-		if((int)s.st_mtime > (int)*mt)
+		if((int)s.st_mtime > (int)(unsigned long)*mt)
 		{
 			ret = 1;
 			*mt = s.st_mtime;
